python_sdk_example/import_export/requirements.txt
boto3
pycryptodome
psec